// src/types/ReviewType/ReviewType.ts
export interface ReviewType {
    review_id: string;
    title: string;
    rating: string;
    content: string;
    visited_date: string;
    privacy: string;
    created_date: string;
    updated_date: string;
    tag_names: string[];
    image_urls: string[];
    valley_name: string;
    valley_id: string;
  }
  